reorganize + build sort function
